Add models to do batched updates
This allows setting multiple properties in one HTTP request to the
nanoleaf device. There's no noticeable difference but it avoids some
issue in theory if one request was lost or delayed (due to a
network blip, process not getting scheduled for some time on low end
hardware, et c) which otherwise would only apply one of the wanted
updates to be applied.

Adds a few utility methods to help define these models easily.